app/pyproject.toml
[project]
name = "cascadeguard-tool"
version = "0.1.0"
description = "CascadeGuard Dockerfile Analysis Tool"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
cascadeguard = "app:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["app", "actions_policy"]
packages = ["scan"]
